src/backend/app/service/webscraping/requirements.txt
selenium
webdriver-manager
requests
PyPDF2
python-dotenv
pika
